put unit test of assistant handler to separate module
