+++ functions/tests/index.spec.js
@@ -3,7 +3,6 @@ const rewire = require('rewire');
 const sinon = require('sinon');
 
 const {buildIntentRequest, MockResponse} = require('./_utils/mocking');
-const {wait} = require('./_utils/wait');
 
 let index, configStub, adminInitStub, functions, admin;
 
@@ -33,22 +32,6 @@ describe('assistant', () => {
       lastSeen: null,
     }), res);
     expect(res.data()).to.have.property('actions').to.have.property('action', 'welcome');
-  });
-
-  it('should warn in case of missed action', () => {
-    const action = 'on-definitely-uncovered-action';
-    let warning = sinon.spy();
-    index.__set__('warning', warning);
-    index.assistant(buildIntentRequest({
-      action,
-      lastSeen: null,
-    }), res);
-
-    return wait()
-      .then(() => {
-        expect(warning.getCall(0).args[0]).to.includes(action);
-        expect(warning).to.be.calledOnce;
-      });
   });
 
   after(() => {
